Process.java: Secondary constructor of Process uses longs instead of ints

+++ src/main/java/com/teragrep/jos_01/procfs/Process.java
@@ -61,7 +61,7 @@ public class Process {
     private final Logger LOGGER = LoggerFactory.getLogger(Process.class);
 
     public Process(String processId) {
-        this(Integer.parseInt(processId));
+        this(Long.parseLong(processId));
     }
 
     public Process(long processId) {
